Rename transitImmutable.js to transitImmutableConverter.js
